changed a folder name from core to shared

+++ include/tradeVolume/tradeVolumeContainer.hpp
@@ -4,7 +4,7 @@
 #include <unordered_map>
 #include <map>
 
-#include "core/marketDataTypes.hpp"
+#include "shared/marketDataTypes.hpp"
 #include "external/simdjson/simdjson.h"
 #include "tradeVolume/tradeVolumeWebSocketEvent.hpp"
 
@@ -13,7 +13,7 @@ namespace MarketData {
 class TradeVolumeContainer {
 private:
     std::mutex mtx_;
-    std::map<int, int> tradeVolume_;
+    std::map<long long int, long long int> tradeVolume_;
 public:
     auto updateFromEvent(TradeVolumeWebSocketEvent& event) -> void;
